Adds the Checkout Open Mode to the catalog definitions
Checkout Open Mode is a metric that indicates the way the user is accessing checkout. At the moment, CHO is in v2, so all modes will have the prefix v2 attached to it. Adding to it, there are two products (checkout and tokenizer) and three modes (redirect, modal and embed).